% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timemachine.R
\name{wktimemachine_version}
\alias{wktimemachine_version}
\title{The wickra-timemachine library version.}
\usage{
wktimemachine_version()
}
\value{
A version string.
}
\description{
The wickra-timemachine library version.
}
